SIMPLY-3091 Refactor AppDelegate and Settings code for SE and OE
App-specific code is now moved to extensions that are individually (mutually-exclusively) built on each app target.
