+++ Simplified/SimplyE-Bridging-Header.h
@@ -1,7 +1,4 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
+#import "NYPLAppDelegate.h"
 #import "NYPLConfiguration.h"
 #import "NYPLBook.h"
 #import "NYPLBookDetailView.h"
